Migrate SyncPatronsJob
