+++ server/src/api/jobs/instances/ScheduleCompetitionEventsJob.ts
@@ -11,7 +11,8 @@ class ScheduleCompetitionEventsJob implements JobDefinition<unknown> {
   async execute() {
     // We're migrating to a new job manager, but jobs for the current one are still in-queue
     // so to prevent duplicating code, just use the old job manager to execute the job on the new one
-    new NewScheduleCompetitionEventsJob().execute();
+    // Once this old job is no longer in use, we can remove this entire file.
+    await new NewScheduleCompetitionEventsJob().execute();
   }
 }
