+++ server/src/api/jobs/job.manager.ts
@@ -63,10 +63,10 @@ const CRON_JOBS = [
   //   type: JobType.SYNC_API_KEYS, // moved to the experimental job manager, for monitoring
   //   interval: '* * * * *' // every 1 min
   // },
-  {
-    type: JobType.SYNC_PATRONS,
-    interval: '* * * * *' // every 1 min
-  },
+  // {
+  //   type: JobType.SYNC_PATRONS,
+  //   interval: '* * * * *' // every 1 min
+  // },
   {
     type: JobType.AUTO_UPDATE_PATRON_PLAYERS,
     interval: '*/5 * * * *' // every 5 mins
